+++ store/example_test.go
@@ -2,6 +2,7 @@ package store_test
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/nelsen129/player-league/store"
 )
@@ -15,7 +16,45 @@ func ExampleInMemoryPlayerStore() {
 	if err != nil {
 		fmt.Println(err)
 	}
+	playerStore.RecordWin("Larry")
+	league := playerStore.GetLeague()
 
 	fmt.Println(score)
-	// Output: 2
+	fmt.Println(league)
+	// Output:
+	// 2
+	// [{Pepper 2} {Larry 1}]
+}
+
+func ExampleFileSystemPlayerStore() {
+	file, err := os.CreateTemp("", "db")
+	if err != nil {
+		fmt.Println(err)
+	}
+	file.Write([]byte(""))
+
+	defer func() {
+		file.Close()
+		os.Remove(file.Name())
+	}()
+
+	playerStore, err := store.NewFileSystemPlayerStore(file)
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	playerStore.RecordWin("Pepper")
+	playerStore.RecordWin("Pepper")
+	score, err := playerStore.GetPlayerScore("Pepper")
+	if err != nil {
+		fmt.Println(err)
+	}
+	playerStore.RecordWin("Larry")
+	league := playerStore.GetLeague()
+
+	fmt.Println(score)
+	fmt.Println(league)
+	// Output:
+	// 2
+	// [{Pepper 2} {Larry 1}]
 }