+++ store/league.go
@@ -9,6 +9,8 @@ import (
 // League type. For finding and sorting players
 type League []Player
 
+// Find returns a pointer to a Player if the player exists
+// in the league, or nil if they don't exist
 func (l League) Find(name string) *Player {
 	for i, p := range l {
 		if p.Name == name {
@@ -30,6 +32,7 @@ func (l League) Swap(i, j int) {
 	l[i], l[j] = l[j], l[i]
 }
 
+// NewLeague parses a JSON string into a League
 func NewLeague(rdr io.Reader) (League, error) {
 	var league League
 	err := json.NewDecoder(rdr).Decode(&league)