+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [Unreleased]
 
 ### Added
-- FileSystemPlayerStore
+- FileSystemPlayerStore, which stores the league database in a JSON file on disc
 ### Changed
 ### Deprecated
 ### Removed